requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
aiohttp==3.12.12
python-multipart==0.0.20
aiofiles==24.1.0
python-dotenv==1.1.0
Requests==2.32.4
requests_negotiate_sspi==0.5.2
requests_ntlm==1.3.0
scikit_learn==1.7.0
cryptography==38.0.4
PyYAML==6.0.2